third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/styles-commit-editing.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function test() {
  var treeElement;
  var treeOutline;
  var section;

  TestRunner.runTestSuite([
    function selectInitialNode(next) {
      ElementsTestRunner.selectNodeAndWaitForStyles('inspected', next);
    },

    function testFreeFlowEdit(next) {
      ElementsTestRunner.dumpSelectedElementStyles(true);
      treeElement = ElementsTestRunner.getElementStylePropertyTreeItem('color');
      treeOutline = treeElement.treeOutline;

      treeElement.startEditing();
      treeElement.nameElement.textContent = 'color';
      treeElement.nameElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));

      // Update incrementally, do not commit.
      treeElement.valueElement.textContent = 'rgb(/*';
      ElementsTestRunner.waitForStyleApplied(next);
      treeElement.kickFreeFlowStyleEditForTest();
    },

    function testCommitEditing(next) {
      // Commit editing.
      treeElement.valueElement.textContent = 'green';
      treeElement.valueElement.firstChild.select();
      TestRunner.addSniffer(Elements.StylePropertiesSection.prototype, '_afterUpdateFinishedForTest', next);
      treeElement.valueElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
    },

    function testNewPropertyEditorIsCreated(next) {
      var blankTreeElement = treeOutline.rootElement().childAt(1);
      if (!UI.isBeingEdited(blankTreeElement.nameElement)) {
        TestRunner.addResult('No new property editor active!');
        TestRunner.completeTest();
        return;
      }

      // Test Styles pane editor looping.
      TestRunner.addSniffer(Elements.StylePropertiesSection.prototype, '_afterUpdateFinishedForTest', next);
      blankTreeElement.nameElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
    },

    function testCycleThroughPropertyEditing(next) {
      if (!UI.isBeingEdited(treeOutline.firstChild().nameElement)) {
        TestRunner.addResult('Original property name editor not active!');
        TestRunner.completeTest();
        return;
      }

      ElementsTestRunner.selectNodeWithId('other', next);
    },

    function testNodeStyles(next) {
      ElementsTestRunner.selectNodeAndWaitForStyles('inspected', next);
    },

    function dumpStyles(next) {
      TestRunner.addResult('After append:');
      ElementsTestRunner.dumpSelectedElementStyles(true);
      next();
    }
  ]);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that editing is canceled properly after incremental editing.
</p>

<div id="inspected" style="color: red">Text</div>
<div id="other"></div>

</body>
</html>
