third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/styles-commit-editing-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that editing is canceled properly after incremental editing.


Running: selectInitialNode

Running: testFreeFlowEdit
[expanded] 
element.style { ()
    color: red;

[expanded] 
div { (user agent stylesheet)
    display: block;


Running: testCommitEditing

Running: testNewPropertyEditorIsCreated

Running: testCycleThroughPropertyEditing

Running: testNodeStyles

Running: dumpStyles
After append:
[expanded] 
element.style { ()
    color: green;

[expanded] 
div { (user agent stylesheet)
    display: block;
